# Ensure the computation stays in single precision